move supertokens files and fix typescript errors
