+++ website/src/api-client/client-hooks.ts
@@ -1,5 +1,5 @@
-import {ApiClient} from '~/app/api-client/ApiClient';
-import {NextJsApiClient} from '~/app/api-client/NextjsApiClient';
+import {ApiClient} from '~/api-client/ApiClient';
+import {NextJsApiClient} from '~/api-client/NextjsApiClient';
 
 export function useApiOnClient() {
   return new ApiClient('http://localhost:4000');
