+++ website/src/config/supertokens/appInfo.ts
@@ -1,7 +1,9 @@
+import {AppInfo} from 'supertokens-node/lib/build/types';
+
 export const appInfo = {
   appName: "AutoScaler",
   apiDomain: "http://localhost:4000",
   websiteDomain: "http://localhost:3000",
   apiBasePath: "/auth",
   websiteBasePath: "/auth",
-};
+} as const satisfies AppInfo;
